# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Service",
        "description": "API by for longnguyen",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/service": {
            "get": {
                "tags": [
                    "Service"
                ],
                "summary": "Get List Service",
                "description": "Display a listing of the resource.",
                "operationId": "getServices",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "An error occurred while Display a listing the service."
                    }
                }
            },
            "post": {
                "tags": [
                    "Service"
                ],
                "summary": "Create Service",
                "description": "Store a newly created resource in storage.",
                "operationId": "storeService",
                "requestBody": {
                    "description": "Create Service",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "service_name",
                                    "service_description",
                                    "service_price",
                                    "service_status"
                                ],
                                "properties": {
                                    "service_name": {
                                        "type": "string",
                                        "example": "Service Name"
                                    },
                                    "service_description": {
                                        "type": "string",
                                        "example": "Service Description"
                                    },
                                    "service_price": {
                                        "type": "number",
                                        "example": "100000"
                                    },
                                    "service_images": {
                                        "type": "file",
                                        "example": ""
                                    },
                                    "service_status": {
                                        "type": "number",
                                        "example": "0"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Invalid file."
                    },
                    "500": {
                        "description": "An error occurred while creating the service."
                    }
                }
            }
        },
        "/api/v1/service/{id}": {
            "delete": {
                "tags": [
                    "Service"
                ],
                "summary": "Delete Service",
                "operationId": "destroyService",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of service to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Service not found"
                    },
                    "500": {
                        "description": "An error occurred while deleting the service."
                    }
                }
            },
            "patch": {
                "tags": [
                    "Service"
                ],
                "summary": "Update Service",
                "description": "Update Service by ID",
                "operationId": "updateService",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of service to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update Service",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "service_name",
                                    "service_description",
                                    "service_images",
                                    "service_status"
                                ],
                                "properties": {
                                    "service_name": {
                                        "type": "string",
                                        "example": "Service Name"
                                    },
                                    "service_description": {
                                        "type": "string",
                                        "example": "Service Description"
                                    },
                                    "service_price": {
                                        "type": "number",
                                        "example": "20000"
                                    },
                                    "service_images": {
                                        "type": "file",
                                        "example": "Service Images"
                                    },
                                    "service_status": {
                                        "type": "number",
                                        "example": "0"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Service not found"
                    },
                    "422": {
                        "description": "Invalid file."
                    },
                    "500": {
                        "description": "An error occurred while updating the service."
                    }
                }
            }
        }
    }
}